+++ macros/generics/activeAuraDamageAndConditionOnEntry.js
@@ -216,7 +216,7 @@ if (args[0].tag === "OnUse" && args[0].macroPass === "preActiveEffects") {
 
   console.debug("ItemMacro: Pre-apply finised, applying effect to template")
 
-  return await AAhelpers.applyTemplate(args);
+  return await game.modules.get("ActiveAuras").api.AAHelpers.applyTemplate(args);
 
 } else if (args[0].tag === "OnUse" && args[0].macroPass === "postActiveEffects") {
   if (lastArg.item.flags.ddbimporter?.effect?.applyImmediate) {